Slide is now switched by clicking.
The current slide is now switched by clicking anywhere in the
slideshow. This fixes a bug where slides would be switched
constantly whenever someone were typing in text in the dashboard.

Catching key events without an input element is currently
unpredictable in meteor which is why the input method was changed.